Added new 2018 data selections associated with the 70MHz instrument.
These are partial results, as new metadata will be uploaded in the
coming weeks. At the moment there is roughly 7 days' worth of
good-quality data associated with the EW polarization of the 70MHz
instrument, and abour a day's worth of good-quality associated with the
NS polarization of the 70MHz instrument.

These selection files will be updated accordingly as new
information is added to the metadatabase.