Revised and polished several functions. Introduced the `Data` class.
Most functions have been revised and polished so as to work with the
latest data loading conventions, as well as to make the code more
uniform.

A new `Data` class has been introduced. This is a firt step which will
hopefully make the code more organized, and the code data objects more customizable.