Added a warning to the `interpolate` method.
If the threshold exceeds the time interval spanned by the data, the
interpolation routine might "wrap around" the data and inadvertently
mistake a distant spectra for a neighboring. This could lead to subtle
unexpected behavior which might be hard to detect.

The `interpolate` method now raises an exception when that scenario is
detected.