Included a pre-generated 'metadatabase.db' binary.
This should help folks hoping to work with the PRIZM metadatabase on the
Niagara SciNet cluster. Indeed, Niagara only makes an older version of
SQLite available to users, making it difficult to run the PRIZM
metadatabase installation scripts. With the a pre-generated binary file
now available as part of the repository, this should no longer be a
problem.

The installation script should still work when executed in a machine
running SQLite vesion 3.36 or superior, however.